add testing framework
