+++ tests.out
@@ -1,0 +1,74 @@
+S: BEAN
+BEAN, INGER, 6, 111, NISTENDIRK, NANCY
+Student: BEAN
+BEAN, INGER, 6, 111, NISTENDIRK, NANCY
+S: ASDF
+S: GUM B
+GUM, NOAH, 54
+S: GUM Bus
+GUM, NOAH, 54
+S: QWERTY Bus
+T: STEIB
+WOOLERY, NOLAN
+VILARDO, EMMANUEL
+Teacher: STEIB
+WOOLERY, NOLAN
+VILARDO, EMMANUEL
+T: BIETS
+G: 1
+SAELEE,DANILO
+GARTH,JOHN
+Grade: 1
+SAELEE,DANILO
+GARTH,JOHN
+G: -1
+B: 56
+RACANELLO, NOEL, 6, 102
+VILARDO, EMMANUEL, 2, 104
+KRAKE, VALENTIN, 6, 109
+BEAN, INGER, 6, 111
+BUSSMANN, BILLY, 2, 108
+FEYEREISEN, WILFREDO, 6, 111
+Bus: 56
+RACANELLO, NOEL, 6, 102
+VILARDO, EMMANUEL, 2, 104
+KRAKE, VALENTIN, 6, 109
+BEAN, INGER, 6, 111
+BUSSMANN, BILLY, 2, 108
+FEYEREISEN, WILFREDO, 6, 111
+B: 100
+G: 1 H
+GARTH,JOHN,3.14,FALKER,ADOLPH,0
+G: 1 High
+GARTH,JOHN,3.14,FALKER,ADOLPH,0
+G: 1 L
+SAELEE,DANILO,2.85,FALKER,ADOLPH,54
+G: 1 Low
+SAELEE,DANILO,2.85,FALKER,ADOLPH,54
+A: 1
+Grade: 1, Avg GPA 2.995
+Average: 1
+Grade: 1, Avg GPA 2.995
+Average: 7
+Grade: 7, Avg GPA 0.0
+I
+0: 0
+1: 2
+2: 13
+3: 9
+4: 15
+5: 0
+6: 21
+Info
+0: 0
+1: 2
+2: 13
+3: 9
+4: 15
+5: 0
+6: 21
+Hello
+Please enter a valid input.
+Q
+
+Process finished with exit code 0
