+++ tests.txt
@@ -1,0 +1,222 @@
+// CSC 365
+// Lab 1-a test suite
+
+// TC-1a
+// Tests Requirements R3, R4
+// short form command name, existing student
+// expected output: BEAN,INGER,6,111,56,2.84,NISTENDIRK,NANCY
+
+S: BEAN
+
+// TC-1b
+// Tests Requirements R3, R4
+// long form command name, existing student
+// expected output: BEAN,INGER,6,111,56,2.84,NISTENDIRK,NANCY
+
+Student: BEAN
+
+// TC-1c
+// Tests Requirements R3, R4
+// short form command name, existing student
+// expected output: <empty line>
+
+S: ASDF
+
+// TC-2a
+// Tests Requirements R3, R5
+// short form command name, existing student, short form Bus
+// expected output: GUM,NOAH,54
+
+S: GUM B
+
+// TC-2b
+// Tests Requirements R3, R5
+// short form command name, existing student, long form Bus
+// expected output: GUM,NOAH,54
+
+S: GUM Bus
+
+// TC-2c
+// Tests Requirements R3, R5
+// short form command name, non-existing student, long form Bus
+// expected output: <empty line>
+
+S: QWERTY Bus
+
+// TC-3a
+// Tests Requirements R3, R6
+// short form command name, existing teacher
+// expected output:
+// WOOLERY,NOLAN
+// VILARDO,EMMANUEL
+
+T: STEIB
+
+// TC-3b
+// Tests Requirements R3, R6
+// long form command name, existing teacher
+// expected output:
+// WOOLERY,NOLAN
+// VILARDO,EMMANUEL
+
+Teacher: STEIB
+
+// TC-3c
+// Tests Requirements R3, R6
+// short form command name, non-existing student
+// expected output: <empty line>
+
+T: BIETS
+
+// TC-4a
+// Tests Requirements R3, R7
+// short form command name, grade
+// expected output:
+// SAELEE,DANILO
+// GARTH,JOHN
+
+G: 1
+
+// TC-4b
+// Tests Requirements R3, R7
+// long form command name, grade
+// expected output:
+// SAELEE,DANILO
+// GARTH,JOHN
+
+Grade: 1
+
+// TC-4c
+// Tests Requirements R3, R7
+// short form command name, non-existent grade
+// expected output: <empty line>
+
+G: -1
+
+// TC-5a
+// Tests Requirements R3, R8
+// short form command name
+// expected output:
+// RACANELLO,NOEL,6,102
+// VILARDO,EMMANUEL,2,104
+// KRAKE,VALENTIN,6,109
+// BEAN,INGER,6,111
+// BUSSMANN,BILLY,2,108
+// FEYEREISEN,WILFREDO,6,111
+
+B: 56
+
+// TC-5b
+// Tests Requirements R3, R8
+// Long form command name
+// expected output:
+// RACANELLO,NOEL,6,102
+// VILARDO,EMMANUEL,2,104
+// KRAKE,VALENTIN,6,109
+// BEAN,INGER,6,111
+// BUSSMANN,BILLY,2,108
+// FEYEREISEN,WILFREDO,6,111
+
+Bus: 56
+
+// TC-5c
+// Tests Requirements R3, R8
+// short form command name, grade
+// expected output: <empty line>
+
+B: 100
+
+// TC-6a
+// Tests Requirements R3, R9
+// short form command name, grade + H
+// expected output:
+// GARTH,JOHN,3.14,FALKER,ADOLPH,0
+
+G: 1 H
+
+// TC-6b
+// Tests Requirements R3, R9
+// short form command name, grade + High
+// expected output:
+// GARTH,JOHN,3.14,FALKER,ADOLPH,0
+
+G: 1 High
+
+// TC-6c
+// Tests Requirements R3, R9
+// short form command name, grade + L
+// expected output:
+// SAELEE,DANILO,2.85,FALKER,ADOLPH, 54
+
+G: 1 L
+
+// TC-6d
+// Tests Requirements R3, R9
+// short form command name, grade + Low
+// expected output:
+// SAELEE,DANILO,2.85,FALKER,ADOLPH, 54
+
+G: 1 Low
+
+// TC-7a
+// Tests Requirements R3, R10
+// short form command name
+// expected output: Grade: 1, Avg GPA 2.995
+
+A: 1
+
+// TC-7b
+// Tests Requirements R3, R10
+// long form command name
+// expected output: Grade: 1, Avg GPA 2.995
+
+Average: 1
+
+// TC-7c
+// Tests Requirements R3, R10
+// long form command name, non-existent grade
+// expected output: Grade: 7, Avg GPA 0
+
+Average: 7
+
+// TC-8a
+// Tests Requirements R3, R11
+// short form command name
+// expected output:
+// 0: 0
+// 1: 2
+// 2: 13
+// 3: 9
+// 4: 15
+// 5: 0
+// 6: 21
+
+I
+
+// TC-8b
+// Tests Requirements R3, R11
+// short form command name
+// expected output:
+// 0: 0
+// 1: 2
+// 2: 13
+// 3: 9
+// 4: 15
+// 5: 0
+// 6: 21
+
+Info
+
+// TC-9
+// Tests Requirements E1
+// invalid command name
+// expected output: Please enter a valid input.
+
+Hello
+
+// TC-10
+// Tests Requirements R3, R12
+// quit command
+// expected output: program terminates
+
+Q
